Parser: Now using a double instead of a float for TaxiNode positions.

+++ .contrib/Source Code/Parser/DB/Types/TaxiNodes.cs
@@ -9,19 +9,19 @@ namespace ATT.DB.Types
     {
         public long ID { get; set; }
         public string Name_lang { get; set; }
-        public float Pos_0 { get; set; }
-        public float Pos_1 { get; set; }
-        public float Pos_2 { get; set; }
-        public float MapOffset_0 { get; set; }
-        public float MapOffset_1 { get; set; }
-        public float FlightMapOffset_0 { get; set; }
-        public float FlightMapOffset_1 { get; set; }
+        public double Pos_0 { get; set; }
+        public double Pos_1 { get; set; }
+        public double Pos_2 { get; set; }
+        public double MapOffset_0 { get; set; }
+        public double MapOffset_1 { get; set; }
+        public double FlightMapOffset_0 { get; set; }
+        public double FlightMapOffset_1 { get; set; }
         public long ContinentID { get; set; }
         public long ConditionID { get; set; }
         public long CharacterBitNumber { get; set; }
         public long Flags { get; set; }
         public long UiTextureKitID { get; set; }
-        public float Facing { get; set; }
+        public double Facing { get; set; }
         public long SpecialIconConditionID { get; set; }
         public long VisibilityConditionID { get; set; }
         public long MountCreatureID_0 { get; set; }
